Refactor remove getter
