+++ src/main/java/racingcar/RandomMoveCondition.java
@@ -3,8 +3,11 @@ package racingcar;
 import camp.nextstep.edu.missionutils.Randoms;
 
 public class RandomMoveCondition implements MoveCondition {
+
+    private static final int standard_number = 4;
+
     @Override
     public boolean canMove() {
-        return Randoms.pickNumberInRange(0, 9) >= 4;
+        return Randoms.pickNumberInRange(0, 9) >= standard_number;
     }
 }
